aider: Change the "Generate" button to be green.
# Aider chat conversation:

USER: change the generate button to be a green color

ASSISTANT: To change the "Generate" button to be green, we need to modify the `web\index.html` file to update the classes applied to the button. We will replace the Tailwind CSS classes for the blue color with the classes for the green color.

Here is the *SEARCH/REPLACE* block:

web\index.html
```html
<<<<<<< SEARCH
<button id="generateButton"
	class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mr-2">Generate</button>
=======
<button id="generateButton"
	class="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded mr-2">Generate</button>
>>>>>>> REPLACE
```