aider: Modify logging middleware to extract IP address without the port.
# Aider chat conversation:

ASSISTANT: To remove the port from the logging output in the `cmd\lotto-numbers\main.go` file, we need to modify the logging middleware to extract the IP address without the port. We can do this by using the `strings.Split` function to split the `RemoteAddr` string on the colon character and take the first part, which is the IP address.

Here is the *SEARCH/REPLACE block* for the necessary change:

cmd\lotto-numbers\main.go
```go
<<<<<<< SEARCH
			// Log the request details with the time taken to serve the page
			log.Printf("Request from %s: %s %s, Duration: %v",
				r.RemoteAddr, r.Method, r.URL.Path, duration)
=======
			// Log the request details with the time taken to serve the page, without the port number
			ipAddress := strings.Split(r.RemoteAddr, ":")[0]
			log.Printf("Request from %s: %s %s, Duration: %v",
				ipAddress, r.Method, r.URL.Path, duration)
>>>>>>> REPLACE
```

Please note that this change assumes that the `RemoteAddr` will always contain an IP address followed by a colon and a port number. If there is a possibility of receiving an IP address without a port, additional checks should be added to handle such cases.